delete h2

+++ jeongyeon/JPA활용_1/h2/src/main/org/h2/command/dml/ScriptBase.java
@@ -1,42 +1,38 @@
 /*
- * Copyright 2004-2019 H2 Group. Multiple-Licensed under the MPL 2.0,
+ * Copyright 2004-2023 H2 Group. Multiple-Licensed under the MPL 2.0,
  * and the EPL 1.0 (https://h2database.com/html/license.html).
  * Initial Developer: H2 Group
  */
 package org.h2.command.dml;
 
-import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
+import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.io.OutputStream;
+import java.nio.charset.Charset;
 import org.h2.api.ErrorCode;
-import org.h2.api.JavaObjectSerializer;
 import org.h2.command.Prepared;
 import org.h2.engine.Constants;
 import org.h2.engine.Database;
-import org.h2.engine.Session;
+import org.h2.engine.SessionLocal;
 import org.h2.engine.SysProperties;
 import org.h2.expression.Expression;
 import org.h2.message.DbException;
-import org.h2.pagestore.db.LobStorageBackend;
 import org.h2.security.SHA256;
-import org.h2.store.DataHandler;
 import org.h2.store.FileStore;
 import org.h2.store.FileStoreInputStream;
 import org.h2.store.FileStoreOutputStream;
 import org.h2.store.fs.FileUtils;
 import org.h2.tools.CompressTool;
 import org.h2.util.IOUtils;
-import org.h2.util.SmallLRUCache;
 import org.h2.util.StringUtils;
-import org.h2.util.TempFileDeleter;
-import org.h2.value.CompareMode;
 
 /**
  * This class is the base for RunScriptCommand and ScriptCommand.
  */
-abstract class ScriptBase extends Prepared implements DataHandler {
+abstract class ScriptBase extends Prepared {
 
     /**
      * The default name of the script file if .zip compression is used.
@@ -49,9 +45,9 @@ abstract class ScriptBase extends Prepared implements DataHandler {
     protected OutputStream out;
 
     /**
-     * The input stream.
+     * The input reader.
      */
-    protected InputStream in;
+    protected BufferedReader reader;
 
     /**
      * The file name (if set).
@@ -66,7 +62,7 @@ abstract class ScriptBase extends Prepared implements DataHandler {
     private FileStore store;
     private String compressionAlgorithm;
 
-    ScriptBase(Session session) {
+    ScriptBase(SessionLocal session) {
         super(session);
     }
 
@@ -108,12 +104,14 @@ abstract class ScriptBase extends Prepared implements DataHandler {
     void deleteStore() {
         String file = getFileName();
         if (file != null) {
-            FileUtils.delete(file);
+            if (FileUtils.isRegularFile(file)) {
+                FileUtils.delete(file);
+            }
         }
     }
 
     private void initStore() {
-        Database db = session.getDatabase();
+        Database db = getDatabase();
         byte[] key = null;
         if (cipher != null && password != null) {
             char[] pass = password.optimize(session).
@@ -136,7 +134,7 @@ abstract class ScriptBase extends Prepared implements DataHandler {
         }
         if (isEncrypted()) {
             initStore();
-            out = new FileStoreOutputStream(store, this, compressionAlgorithm);
+            out = new FileStoreOutputStream(store, compressionAlgorithm);
             // always use a big buffer, otherwise end-of-block is written a lot
             out = new BufferedOutputStream(out, Constants.IO_BUFFER_SIZE_COMPRESS);
         } else {
@@ -153,28 +151,30 @@ abstract class ScriptBase extends Prepared implements DataHandler {
 
     /**
      * Open the input stream.
+     *
+     * @param charset the charset to use
      */
-    void openInput() {
+    void openInput(Charset charset) {
         String file = getFileName();
         if (file == null) {
             return;
         }
+        InputStream in;
         if (isEncrypted()) {
             initStore();
-            in = new FileStoreInputStream(store, this, compressionAlgorithm != null, false);
+            in = new FileStoreInputStream(store, compressionAlgorithm != null, false);
         } else {
-            InputStream inStream;
             try {
-                inStream = FileUtils.newInputStream(file);
+                in = FileUtils.newInputStream(file);
             } catch (IOException e) {
                 throw DbException.convertIOException(e, file);
             }
-            in = new BufferedInputStream(inStream, Constants.IO_BUFFER_SIZE);
             in = CompressTool.wrapInputStream(in, compressionAlgorithm, SCRIPT_SQL);
             if (in == null) {
                 throw DbException.get(ErrorCode.FILE_NOT_FOUND_1, SCRIPT_SQL + " in " + file);
             }
         }
+        reader = new BufferedReader(new InputStreamReader(in, charset), Constants.IO_BUFFER_SIZE);
     }
 
     /**
@@ -183,8 +183,8 @@ abstract class ScriptBase extends Prepared implements DataHandler {
     void closeIO() {
         IOUtils.closeSilently(out);
         out = null;
-        IOUtils.closeSilently(in);
-        in = null;
+        IOUtils.closeSilently(reader);
+        reader = null;
         if (store != null) {
             store.closeSilently();
             store = null;
@@ -196,73 +196,8 @@ abstract class ScriptBase extends Prepared implements DataHandler {
         return false;
     }
 
-    @Override
-    public String getDatabasePath() {
-        return null;
-    }
-
-    @Override
-    public FileStore openFile(String name, String mode, boolean mustExist) {
-        return null;
-    }
-
-    @Override
-    public void checkPowerOff() {
-        session.getDatabase().checkPowerOff();
-    }
-
-    @Override
-    public void checkWritingAllowed() {
-        session.getDatabase().checkWritingAllowed();
-    }
-
-    @Override
-    public int getMaxLengthInplaceLob() {
-        return session.getDatabase().getMaxLengthInplaceLob();
-    }
-
-    @Override
-    public TempFileDeleter getTempFileDeleter() {
-        return session.getDatabase().getTempFileDeleter();
-    }
-
-    @Override
-    public String getLobCompressionAlgorithm(int type) {
-        return session.getDatabase().getLobCompressionAlgorithm(type);
-    }
-
     public void setCompressionAlgorithm(String algorithm) {
         this.compressionAlgorithm = algorithm;
     }
 
-    @Override
-    public Object getLobSyncObject() {
-        return this;
-    }
-
-    @Override
-    public SmallLRUCache<String, String[]> getLobFileListCache() {
-        return null;
-    }
-
-    @Override
-    public LobStorageBackend getLobStorage() {
-        return null;
-    }
-
-    @Override
-    public int readLob(long lobId, byte[] hmac, long offset, byte[] buff,
-            int off, int length) {
-        throw DbException.throwInternalError();
-    }
-
-    @Override
-    public JavaObjectSerializer getJavaObjectSerializer() {
-        return session.getDataHandler().getJavaObjectSerializer();
-    }
-
-    @Override
-    public CompareMode getCompareMode() {
-        return session.getDataHandler().getCompareMode();
-    }
 }
